Add support for inline comments
